pyproject.toml
[project]
name = "chatbot-gsantana"
version = "0.1.0"
description = "Chat bot para interação de usuários com um site."
authors = [
    {name = "Gsantana",email = "gilmar.jesus@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "fastapi (>=0.111.0,<0.112.0)",
    "uvicorn[standard] (>=0.30.1,<0.31.0)",
    "pydantic-settings (>=2.4.0,<3.0.0)",
    "python-multipart (>=0.0.9,<0.0.10)",
    "requests (>=2.32.4,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "nltk (>=3.9.1,<4.0.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "psycopg2-binary (>=2.9.9,<3.0.0)",
    "python-jose[cryptography] (>=3.3.0,<4.0.0)",
    "argon2-cffi (>=23.1.0,<24.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "bcrypt (<4.0.0)",
    "structlog (>=25.4.0,<26.0.0)",
]

[tool.poetry]
packages = [{include = "chatbot_gsantana", from = "src"}]

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
black = "^25.1.0"
isort = "^6.0.1"
flake8 = "^7.3.0"
pre-commit = "^4.3.0"
pytest-mock = "^3.14.1"
diagrams = "^0.24.4"
pytest-dotenv = "^0.5.2"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.flake8]
max-line-length = 88
extend-ignore = "E203,E501"
